+++ src/org/store.js
@@ -40,9 +40,13 @@ class OrgStore {
     fs.mkdirSync(this._dir, { recursive: true });
     try {
       this._db = new Database(this.dbPath);
-    } catch {
-      try { fs.unlinkSync(this.dbPath); } catch {}
-      this._db = new Database(this.dbPath);
+    } catch (err) {
+      if (err && (err.name === 'SqliteError' || (err.code && String(err.code).startsWith('SQLITE_')))) {
+        try { fs.unlinkSync(this.dbPath); } catch {}
+        this._db = new Database(this.dbPath);
+      } else {
+        throw err;
+      }
     }
     this._db.pragma('journal_mode = WAL');
     this._db.pragma('synchronous = NORMAL');
